+++ packages/medical-domain/src/domains/models/statistics/carbs-statistics.model.ts
@@ -29,8 +29,11 @@ import { HoursRange } from './satistics.model'
 
 export interface CarbsStatistics {
   totalCarbsPerDay: number
-  foodCarbsPerDay: number
-  totalEntriesCarbWithRescueCarbs: number
+  rescueCarbsPerDay: number
+  mealCarbsPerDay: number
+  estimatedCarbsPerDay: number
+  totalMealCarbsWithRescueCarbsEntries: number
+  totalRescueCarbsEntries: number
 }
 
 export interface RescueCarbsAveragePerRange {